Support For Deploying Standalone Processors (#86)
This builds on #84. It adds the functionality to allow for the download and installation of a custom processor using the ProcessorURL field.

The processor will be saved to the volume used by the conduit pod.

Right now, when the pod is recreated it will download the processor again. There is currently no checking of checksums, versions, etc if any are saved and will overwrite the existing processor, given the name being the same.